+++ DataAccess/Util/CategoryMapper.cs
@@ -1,0 +1,28 @@
+using BusinessObjects.Model;
+using DataAccess.DTO;
+
+namespace DataAccess.Util
+{
+    public class CategoryMapper
+    {
+        public static CategoryDTO mapToDTO(Category category)
+        {
+            CategoryDTO categoryDTO = new CategoryDTO
+            {
+                CategoryId = category.CategoryId,
+                CategoryName = category.Name
+            };
+            return categoryDTO;
+        }
+
+        public static Category mapToEntity(CategoryDTO categoryDTO)
+        {
+            Category category = new Category
+            {
+                CategoryId = categoryDTO.CategoryId,
+                Name = categoryDTO.CategoryName
+            };
+            return category;
+        }
+    }
+}
