+++ DataAccess/Util/Mapper.cs
@@ -1,37 +1,10 @@
 ﻿using BusinessObjects.Model;
 using DataAccess.DTO;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace DataAccess.Util
 {
     public class Mapper
     {
-        // public static UserDTO? mapToDTO(User user)
-        // {
-        //     if (user != null)
-        //     {
-        //         return new UserDTO
-        //         {
-        //             UserId = user.UserId,
-        //             Email = user.Email,
-        //             Country = user.Country,
-        //             CompanyName = user.CompanyName,
-        //             City = user.City,
-        //             Password = user.Password,
-        //             Role = user.Role
-        //         };
-        //     }
-        //     else
-        //     {
-        //         return null;
-        //     }
-
-        // }
-
         public static OrderDTO mapToDTO(Order order)
         {
             OrderDTO orderDTO = new OrderDTO
@@ -57,17 +30,9 @@ namespace DataAccess.Util
             return productDTO;
         }
 
-        public static CategoryDTO mapToDTO(Category category)
-        {
-            CategoryDTO categoryDTO = new CategoryDTO
-            {
-                CategoryId = category.CategoryId,
-                CategoryName = category.Name
-            };
-            return categoryDTO;
-        }
 
-        
+
+
 
         public static Order mapToEntity(OrderDTO orderDTO)
         {
@@ -93,14 +58,6 @@ namespace DataAccess.Util
             return product;
         }
 
-        public static Category mapToEntity(CategoryDTO categoryDTO)
-        {
-            Category category = new Category
-            {
-                CategoryId = categoryDTO.CategoryId,
-                Name = categoryDTO.CategoryName
-            };
-            return category;
-        }
+
     }
 }
